# MovieVerse-Mobile/platforms/ios/www/MovieVerse-Backend/APIs/movieverse-openapi.yaml
openapi: 3.0.3
info:
  title: MovieVerse Gateway API
  description: Unified edge API for MovieVerse microservices.
  version: 2.0.0
servers:
  - url: http://localhost:8080
    description: Local gateway (Nginx)
paths:
  /movies:
    get:
      summary: List movies
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            default: 50
        - in: query
          name: genre
          schema:
            type: string
      responses:
        "200":
          description: Movie list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Movie"
    post:
      summary: Create a movie
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MovieCreate"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Movie"
  /movies/{movie_id}:
    get:
      summary: Get a movie
      parameters:
        - in: path
          name: movie_id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Movie detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Movie"
    put:
      summary: Update a movie
      parameters:
        - in: path
          name: movie_id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MovieUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Movie"
  /genres:
    get:
      summary: List genres
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            default: 50
        - in: query
          name: q
          schema:
            type: string
      responses:
        "200":
          description: Genre list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenreList"
  /genres/{genre_id}:
    get:
      summary: Get genre
      parameters:
        - in: path
          name: genre_id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Genre detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Genre"
  /people:
    get:
      summary: List people
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            default: 50
        - in: query
          name: q
          schema:
            type: string
        - in: query
          name: department
          schema:
            type: string
      responses:
        "200":
          description: People list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PersonList"
  /people/{person_id}:
    get:
      summary: Get person
      parameters:
        - in: path
          name: person_id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Person detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Person"
  /reviews:
    get:
      summary: List reviews
      parameters:
        - in: query
          name: movie_id
          schema:
            type: integer
        - in: query
          name: user_id
          schema:
            type: integer
      responses:
        "200":
          description: Review list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"
    post:
      summary: Create review
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewCreate"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
  /profiles:
    post:
      summary: Create profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileCreate"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
  /profiles/{user_id}:
    get:
      summary: Get profile
      parameters:
        - in: path
          name: user_id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
    put:
      summary: Update profile
      parameters:
        - in: path
          name: user_id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
  /recommendations:
    post:
      summary: Get recommendations
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecommendationRequest"
      responses:
        "200":
          description: Recommendation list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecommendationResponse"
  /recommendations/similar:
    post:
      summary: Get similar movies
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SimilarRequest"
      responses:
        "200":
          description: Similar list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SimilarResponse"
  /search:
    post:
      summary: Search movies
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SearchRequest"
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SearchResponse"
  /search/reviews:
    post:
      summary: Search reviews
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SearchRequest"
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SearchResponse"
  /notifications:
    get:
      summary: List notifications
      parameters:
        - in: query
          name: user_id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Notification list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notification"
    post:
      summary: Send notification
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NotificationCreate"
      responses:
        "201":
          description: Notification queued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
  /auth/register:
    post:
      summary: Register user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthRegister"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
  /auth/login:
    post:
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthLogin"
      responses:
        "200":
          description: Tokens
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthTokens"
  /auth/refresh:
    post:
      summary: Refresh token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthRefresh"
      responses:
        "200":
          description: Tokens
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthTokens"
  /crawl:
    post:
      summary: Enqueue crawler job
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CrawlRequest"
      responses:
        "202":
          description: Job accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CrawlResponse"
components:
  schemas:
    Movie:
      type: object
      properties:
        movie_id:
          type: integer
        tmdb_id:
          type: integer
          nullable: true
        title:
          type: string
        overview:
          type: string
          nullable: true
        genres:
          type: string
          nullable: true
        release_date:
          type: string
          nullable: true
        rating:
          type: number
          nullable: true
        popularity:
          type: number
          nullable: true
    MovieCreate:
      type: object
      required: [title]
      properties:
        tmdb_id:
          type: integer
        title:
          type: string
        overview:
          type: string
        genres:
          type: string
        release_date:
          type: string
        rating:
          type: number
        popularity:
          type: number
    MovieUpdate:
      type: object
      properties:
        tmdb_id:
          type: integer
        title:
          type: string
        overview:
          type: string
        genres:
          type: string
        release_date:
          type: string
        rating:
          type: number
        popularity:
          type: number
    Genre:
      type: object
      properties:
        genre_id:
          type: integer
        name:
          type: string
    GenreList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Genre"
    Person:
      type: object
      properties:
        person_id:
          type: integer
        name:
          type: string
        biography:
          type: string
          nullable: true
        birthday:
          type: string
          nullable: true
        deathday:
          type: string
          nullable: true
        profile_path:
          type: string
          nullable: true
        known_for_department:
          type: string
          nullable: true
    PersonList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Person"
    Review:
      type: object
      properties:
        review_id:
          type: integer
        user_id:
          type: integer
        movie_id:
          type: integer
        rating:
          type: number
        review_text:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
    ReviewCreate:
      type: object
      required: [user_id, movie_id, rating]
      properties:
        user_id:
          type: integer
        movie_id:
          type: integer
        rating:
          type: number
        review_text:
          type: string
    Profile:
      type: object
      properties:
        user_id:
          type: integer
        display_name:
          type: string
        bio:
          type: string
          nullable: true
        favorite_genres:
          type: array
          items:
            type: string
    ProfileCreate:
      type: object
      required: [user_id, display_name]
      properties:
        user_id:
          type: integer
        display_name:
          type: string
        bio:
          type: string
        favorite_genres:
          type: array
          items:
            type: string
    ProfileUpdate:
      type: object
      properties:
        display_name:
          type: string
        bio:
          type: string
        favorite_genres:
          type: array
          items:
            type: string
    RecommendationRequest:
      type: object
      required: [user_id]
      properties:
        user_id:
          type: integer
        limit:
          type: integer
        filter_genres:
          type: array
          items:
            type: string
    RecommendationResponse:
      type: object
      properties:
        user_id:
          type: integer
        items:
          type: array
          items:
            type: object
            properties:
              movie_id:
                type: integer
              score:
                type: number
    SimilarRequest:
      type: object
      required: [movie_id]
      properties:
        movie_id:
          type: integer
        limit:
          type: integer
    SimilarResponse:
      type: object
      properties:
        movie_id:
          type: integer
        items:
          type: array
          items:
            type: object
            properties:
              movie_id:
                type: integer
              score:
                type: number
    SearchRequest:
      type: object
      required: [query]
      properties:
        query:
          type: string
        limit:
          type: integer
    SearchResponse:
      type: object
      properties:
        results:
          type: array
          items:
            type: object
    Notification:
      type: object
      properties:
        notification_id:
          type: integer
        user_id:
          type: integer
        message:
          type: string
        status:
          type: string
        created_at:
          type: string
          format: date-time
    NotificationCreate:
      type: object
      required: [user_id, message]
      properties:
        user_id:
          type: integer
        message:
          type: string
    AuthRegister:
      type: object
      required: [email, username, password]
      properties:
        email:
          type: string
        username:
          type: string
        password:
          type: string
          format: password
    AuthLogin:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
          format: password
    AuthRefresh:
      type: object
      required: [user_id, refresh_token]
      properties:
        user_id:
          type: integer
        refresh_token:
          type: string
    AuthUser:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        username:
          type: string
    AuthTokens:
      type: object
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
    CrawlRequest:
      type: object
      required: [url]
      properties:
        url:
          type: string
        source:
          type: string
        tags:
          type: array
          items:
            type: string
        priority:
          type: integer
    CrawlResponse:
      type: object
      properties:
        job_id:
          type: string
        url:
          type: string
        status:
          type: string