# MovieVerse-Mobile/www/MovieVerse-Backend/services/movieverse_services/notification_service/models.py
from __future__ import annotations

from sqlalchemy import Column, DateTime, Integer, String, func

from movieverse_services.common.db import Base


class Notification(Base):
    __tablename__ = "notifications"

    notification_id = Column(Integer, primary_key=True, autoincrement=True)
    user_id = Column(Integer, nullable=False, index=True)
    message = Column(String(500), nullable=False)
    created_at = Column(DateTime(timezone=True), server_default=func.now(), nullable=False)
    status = Column(String(20), default="queued", nullable=False)